requirements.txt
backtrader==1.9.78.123
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
yfinance>=0.2.0
requests>=2.25.0
cryptography>=3.4.0
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=1.0.0
google-adk>=1.0.0
python-dotenv>=1.0.0
